# swagger.yaml
openapi: 3.0.0
info:
  title: Ecommerce API
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:3000/api/v1
  variables: {}
paths:
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The user ID
    get:
      tags:
      - Users
      summary: Get Single User
      operationId: GetSingleUser
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /product/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The Product ID
    get:
      tags:
      - Product
      summary: Get Single Product
      operationId: GetSingleProduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    patch:
      tags:
      - Product
      summary: Update Product
      operationId: UpdateProduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Product
      summary: Delete Product
      operationId: DeleteProduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /orders:
    get:
      tags:
      - Order
      summary: Get All Orders
      operationId: GetAllOrders
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    post:
      tags:
      - Order
      summary: Create Order
      operationId: CreateOrder
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateOrderRequest'
              - example:
                  tax: 399
                  shippingFee: 499
                  items:
                  - name: accent chair
                    price: 2599
                    image: https://dl.airtable.com/.attachmentThumbnails/e8bc3791196535af65f40e36993b9e1f/438bd160
                    amount: 10
                    product: 658a7f971da6366194a31d00
            example:
              tax: 399
              shippingFee: 499
              items:
              - name: accent chair
                price: 2599
                image: https://dl.airtable.com/.attachmentThumbnails/e8bc3791196535af65f40e36993b9e1f/438bd160
                amount: 10
                product: 658a7f971da6366194a31d00
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /orders/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
        description: The Product ID
    get:
      tags:
      - Order
      summary: get Single Order
      operationId: getSingleOrder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    patch:
      tags:
      - Order
      summary: update Single Order
      operationId: updateSingleOrder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
components:
  schemas:
...
    CreateOrderRequest:
      title: CreateOrderRequest
      required:
      - tax
      - shippingFee
      - items
      type: object
      properties:
        tax:
          type: integer
          format: int32
        shippingFee:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
          description: ''
      example:
        tax: 399
        shippingFee: 499
        items:
        - name: accent chair
          price: 2599
          image: https://dl.airtable.com/.attachmentThumbnails/e8bc3791196535af65f40e36993b9e1f/438bd160
          amount: 10
          product: 658a7f971da6366194a31d00
    Item:
      title: Item
      required:
      - name
      - price
      - image
      - amount
      - product
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          format: int32
        image:
          type: string
        amount:
          type: integer
          format: int32
        product:
          type: string
      example:
        name: accent chair
        price: 2599
        image: https://dl.airtable.com/.attachmentThumbnails/e8bc3791196535af65f40e36993b9e1f/438bd160
        amount: 10
        product: 658a7f971da6366194a31d00
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...